requirements.txt
streamlit
openai
mysql.connector
toml
mysql-connector-python